requirements.txt
certifi==2018.1.18
Django==2.0.2
django-background-tasks==1.1.13
django-compat==1.0.15
pytz==2018.3
requests==2.18.4
